+++ ixmp/backend/jdbc.py
@@ -407,30 +407,55 @@ class JDBCBackend(CachingBackend):
 
     # Timeseries methods
 
-    def _common_init(self, ts, klass, *args):
-        """Common code for ts_init and s_init."""
-        method = getattr(self.jobj, 'new' + klass)
-        # Create a new TimeSeries
-        jobj = method(ts.model, ts.scenario, *args)
+    def _index_and_set_attrs(self, jobj, ts):
+        """Add *jobj* to index and update attributes of *ts*.
 
+        Helper for init and get.
+        """
         # Add to index
         self.jindex[ts] = jobj
 
-        # Retrieve initial version
-        ts.version = jobj.getVersion()
+        # Retrieve the version of the Java object
+        v = jobj.getVersion()
+        if ts.version is None:
+            # The default version was requested; update the attribute
+            ts.version = v
+        elif v != ts.version:  # pragma: no cover
+            # Something went wrong on the Java side
+            raise RuntimeError(f'got version {v} instead of {ts.version}')
 
-    def init_ts(self, ts, annotation=None):
-        self._common_init(ts, 'TimeSeries', annotation)
+        if isinstance(ts, Scenario):
+            # Also retrieve the scheme
+            s = jobj.getScheme()
 
-    def get(self, ts, version):
+            if ts.scheme and s != ts.scheme:  # pragma: no cover
+                # Something went wrong on the Java side
+                raise RuntimeError(f'got scheme {s} instead of {ts.scheme}')
+
+            ts.scheme = s
+
+    def init(self, ts, annotation):
+        klass = ts.__class__.__name__
+
+        # Final arguments: scheme only for Scenarios
+        args = [ts.scheme, annotation] if klass == 'Scenario' else [annotation]
+
+        # Call either newTimeSeries or newScenario
+        method = getattr(self.jobj, 'new' + klass)
+        jobj = method(ts.model, ts.scenario, *args)
+
+        self._index_and_set_attrs(jobj, ts)
+
+    def get(self, ts):
         args = [ts.model, ts.scenario]
-        if version is not None:
+        if ts.version is not None:
             # Load a TimeSeries of specific version
-            args.append(version)
+            args.append(ts.version)
 
         # either getTimeSeries or getScenario
         method = getattr(self.jobj, 'get' + ts.__class__.__name__)
         try:
+            # Either the 2- or 3- argument form, depending on args
             jobj = method(*args)
         except java.IxException as e:
             # Try to re-raise as a ValueError for bad model or scenario name
@@ -443,21 +468,7 @@ class JDBCBackend(CachingBackend):
             # Failed
             _raise_jexception(e)
 
-        # Add to index
-        self.jindex[ts] = jobj
-
-        # Retrieve or check the version
-        if version is None:
-            version = jobj.getVersion()
-        else:
-            assert version == jobj.getVersion()
-
-        # Update the version attribute
-        ts.version = version
-
-        if isinstance(ts, Scenario):
-            # Also retrieve the scheme
-            ts.scheme = jobj.getScheme()
+        self._index_and_set_attrs(jobj, ts)
 
     def check_out(self, ts, timeseries_only):
         try:
@@ -581,10 +592,6 @@ class JDBCBackend(CachingBackend):
         self.jindex[ts].removeGeoData(region, variable, subannual, years, unit)
 
     # Scenario methods
-
-    def init_s(self, s, scheme, annotation):
-        self._common_init(s, 'Scenario', scheme, annotation)
-
     def clone(self, s, platform_dest, model, scenario, annotation,
               keep_solution, first_model_year=None):
         # Raise exceptions for limitations of JDBCBackend
@@ -611,7 +618,8 @@ class JDBCBackend(CachingBackend):
 
         # Instantiate same class as the original object
         return s.__class__(platform_dest, model, scenario,
-                           version=jclone.getVersion())
+                           version=jclone.getVersion(),
+                           scheme=jclone.getScheme())
 
     def has_solution(self, s):
         return self.jindex[s].hasSolution()