+++ ixmp/tests/reporting/test_computations.py
@@ -11,9 +11,6 @@ from ixmp.testing import assert_logs
 from . import add_test_data
 
 
-REGISTRY = pint.get_application_registry()
-
-
 @pytest.fixture(scope='function')
 def data(test_mp, request):
     scen = ixmp.Scenario(test_mp, request.node.name, request.node.name, 'new')
@@ -25,10 +22,12 @@ def test_apply_units(data, caplog):
     # Unpack
     *_, x = data
 
+    registry = pint.get_application_registry()
+
     # Brute-force replacement with incompatible units
     with assert_logs(caplog, "Replace 'kilogram' with incompatible 'liter'"):
         result = computations.apply_units(x, 'litres')
-    assert result.attrs['_unit'] == REGISTRY.Unit('litre')
+    assert result.attrs['_unit'] == registry.Unit('litre')
     # No change in values
     assert_series_equal(result.to_series(), x.to_series())
 
@@ -37,17 +36,17 @@ def test_apply_units(data, caplog):
     # Compatible units: magnitudes are also converted
     with assert_logs(caplog, "Convert 'kilogram' to 'metric_ton'"):
         result = computations.apply_units(x, 'tonne')
-    assert result.attrs['_unit'] == REGISTRY.Unit('tonne')
+    assert result.attrs['_unit'] == registry.Unit('tonne')
     assert_series_equal(result.to_series(), x.to_series() * 0.001)
 
     # Remove unit
-    x.attrs['_unit'] = REGISTRY.Unit('dimensionless')
+    x.attrs['_unit'] = registry.Unit('dimensionless')
 
     caplog.clear()
     result = computations.apply_units(x, 'kg')
     # Nothing logged when _unit attr is missing
     assert len(caplog.messages) == 0
-    assert result.attrs['_unit'] == REGISTRY.Unit('kg')
+    assert result.attrs['_unit'] == registry.Unit('kg')
     assert_series_equal(result.to_series(), x.to_series())